+++ languagetool-language-modules/fa/src/main/java/org/languagetool/rules/fa/PersianSpaceBeforeRule.java
@@ -18,29 +18,28 @@
  */
 package org.languagetool.rules.fa;
 
-import java.util.ArrayList;
-import java.util.List;
 import java.util.ResourceBundle;
-import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
-import org.languagetool.AnalyzedSentence;
-import org.languagetool.AnalyzedTokenReadings;
 import org.languagetool.Language;
-import org.languagetool.rules.Category;
-import org.languagetool.rules.RuleMatch;
+import org.languagetool.rules.AbstractSpaceBeforeRule;
 
 /**
- * A Persian rule that checks if there is a missing space begore some conjunctions.
+ * A Persian rule that checks if there is a missing space before some conjunctions.
  * 
  * @since 2.7
  */
-public class PersianSpaceBeforeRule extends PersianRule {
+public class PersianSpaceBeforeRule extends AbstractSpaceBeforeRule {
 
-  private static final Pattern CONJUNCTIONS = Pattern.compile("و|به|با|تا|زیرا|چون|بنابراین|چونکه",Pattern.CASE_INSENSITIVE|Pattern.UNICODE_CASE);
-   
-  public PersianSpaceBeforeRule(final ResourceBundle messages, final Language language) {
-    super.setCategory(new Category(messages.getString("category_misc")));
+  public PersianSpaceBeforeRule(ResourceBundle messages, Language language) {
+    super(messages, language);
+  }
+
+  private static final Pattern CONJUNCTIONS = Pattern.compile("و|به|با|تا|زیرا|چون|بنابراین|چونکه");
+
+  @Override
+  protected Pattern getConjuntions() {
+    return CONJUNCTIONS;
   }
 
   @Override
@@ -54,38 +53,13 @@ public class PersianSpaceBeforeRule extends PersianRule {
   }
 
   @Override
-  public final RuleMatch[] match(final AnalyzedSentence sentence) {
-    final List<RuleMatch> ruleMatches = new ArrayList<>();
-    final AnalyzedTokenReadings[] tokens = sentence.getTokens();
-    
-    for (int i = 1; i < tokens.length; i++) {
-      final String token = tokens[i].getToken();
-      Matcher matcher = CONJUNCTIONS.matcher(token);
-      if (matcher.matches()) {
-        if (!tokens[i - 1].getToken().equals(" ")) {
-          final String replacement = " " + token;
-          final String msg = getSuggestion();
-          final int pos = tokens[i].getStartPos();
-          final RuleMatch potentialRuleMatch = new RuleMatch(this, pos, pos
-              + token.length(), msg, getShort());
-          potentialRuleMatch.setSuggestedReplacement(replacement);
-          ruleMatches.add(potentialRuleMatch);
-        }
-      }
-    }
-    return toRuleMatchArray(ruleMatches);
-  }
-
-  private String getShort() {
+  protected String getShort() {
     return "فاصلهٔ حذف‌شده";
   }
 
-  private String getSuggestion() {
+  @Override
+  protected String getSuggestion() {
     return "فاصلهٔ قبل از حرف ربط حذف شده‌است";
   }
-
-  @Override
-  public void reset() {
-  }  
 
 }
