+++ languagetool-language-modules/km/src/main/java/org/languagetool/rules/km/KhmerSpaceBeforeRule.java
@@ -27,65 +27,30 @@ import java.util.regex.Pattern;
 import org.languagetool.AnalyzedSentence;
 import org.languagetool.AnalyzedTokenReadings;
 import org.languagetool.Language;
+import org.languagetool.rules.AbstractSpaceBeforeRule;
 import org.languagetool.rules.Category;
 import org.languagetool.rules.RuleMatch;
 
 /**
- * A Khmer rule that checks if there is a missing space begore some conjunctions.
+ * A Khmer rule that checks if there is a missing space before some conjunctions.
  * 
  * @author Jaume Ortolà
  */
-public class KhmerSpaceBeforeRule extends KhmerRule {
+public class KhmerSpaceBeforeRule extends AbstractSpaceBeforeRule {
+
+  public KhmerSpaceBeforeRule(final ResourceBundle messages, final Language language) {
+    super(messages, language);
+  }
 
   private static final Pattern CONJUNCTIONS = Pattern.compile("ដើម្បី|និង|ពីព្រោះ",Pattern.CASE_INSENSITIVE|Pattern.UNICODE_CASE);
-   
-  public KhmerSpaceBeforeRule(final ResourceBundle messages, final Language language) {
-    super.setCategory(new Category(messages.getString("category_misc")));
+
+  @Override
+  public Pattern getConjuntions() {
+    return CONJUNCTIONS;
   }
 
   @Override
   public String getId() {
     return "KM_SPACE_BEFORE_CONJUNCTION";
   }
-
-  @Override
-  public String getDescription() {
-    return "Checks for missing space before some conjunctions";
-  }
-
-  @Override
-  public final RuleMatch[] match(final AnalyzedSentence sentence) {
-    final List<RuleMatch> ruleMatches = new ArrayList<>();
-    final AnalyzedTokenReadings[] tokens = sentence.getTokens();
-    
-    for (int i = 1; i < tokens.length; i++) {
-      final String token = tokens[i].getToken();
-      Matcher matcher = CONJUNCTIONS.matcher(token);
-      if (matcher.matches()) {
-        if (!tokens[i - 1].getToken().equals(" ")) {
-          final String replacement = " " + token;
-          final String msg = getSuggestion();
-          final int pos = tokens[i].getStartPos();
-          final RuleMatch potentialRuleMatch = new RuleMatch(this, pos, pos
-              + token.length(), msg, getShort());
-          potentialRuleMatch.setSuggestedReplacement(replacement);
-          ruleMatches.add(potentialRuleMatch);
-        }
-      }
-    }
-    return toRuleMatchArray(ruleMatches);
-  }
-
-  private String getShort() {
-    return "Missing white space";
-  }
-
-  private String getSuggestion() {
-    return "Missing white space before conjunction";
-  }
-
-  @Override
-  public void reset() {
-  }  
-
 }
